move systemlib/airspeed to standalone lib
