+++ src/drivers/differential_pressure/ets/ets_airspeed.cpp
@@ -38,11 +38,12 @@
  * Driver for the Eagle Tree Airspeed V3 connected via I2C.
  */
 
+#include <cfloat>
+
 #include <px4_config.h>
 
 #include <drivers/device/i2c.h>
 
-#include <systemlib/airspeed.h>
 #include <systemlib/err.h>
 #include <parameters/param.h>
 #include <perf/perf_counter.h>
