+++ src/drivers/differential_pressure/ms4525/ms4525_airspeed.cpp
@@ -54,7 +54,6 @@
 
 #include <drivers/device/i2c.h>
 
-#include <systemlib/airspeed.h>
 #include <systemlib/err.h>
 #include <parameters/param.h>
 #include <perf/perf_counter.h>
