+++ src/modules/mavlink/CMakeLists.txt
@@ -61,6 +61,7 @@ px4_add_module(
 		mavlink_ulog.cpp
 		mavlink_timesync.cpp
 	DEPENDS
+		airspeed
 		git_mavlink_v2
 		conversion
 		git_ecl
