+++ src/modules/simulator/airspeedsim/airspeedsim.cpp
@@ -54,7 +54,6 @@
 #include <math.h>
 #include <unistd.h>
 
-#include <systemlib/airspeed.h>
 #include <systemlib/err.h>
 #include <parameters/param.h>
 #include <perf/perf_counter.h>
