+++ src/modules/simulator/airspeedsim/meas_airspeed_sim.cpp
@@ -59,7 +59,6 @@
 #include <math.h>
 #include <unistd.h>
 
-#include <systemlib/airspeed.h>
 #include <systemlib/err.h>
 #include <parameters/param.h>
 #include <perf/perf_counter.h>
